feat(utils): extract color to utils & fix submenu color
